take out logic from app to hoc
