+++ package.json
@@ -34,6 +34,7 @@
     "react": "16.8.6",
     "react-dom": "16.8.6",
     "react-redux": "7.0.3",
+    "recompose": "0.30.0",
     "redux": "4.0.1"
   },
   "jest": {
