+++ src/components/app/app.jsx
@@ -1,34 +1,7 @@
-import {connect} from 'react-redux';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import {ActionCreator} from '../../reducer';
-import {ArtistQuestionScreen} from '../artist-question-screen/artist-question-screen.jsx';
-import {GenreQuestionScreen} from '../genre-question-screen/genre-question-screen.jsx';
-import {WelcomeScreen} from '../welcome-screen/welcome-screen.jsx';
-import LosingScene from '../losing-scene/losing-scene.jsx';
 import MistakeScreen from '../mistakeScreen/mistakeScreen.jsx';
-import VictoryScene from '../victory-scene/victory-scene.jsx';
-
-import withActivePlayer from '../../hocs/with-active-player/with-active-player';
-import withTransformProps from '../../hocs/with-transform-props/with-transform-props';
-import withUserAnswer from '../../hocs/with-user-answer/with-user-answer';
-
-const transformPlayerToAnswer = (props) => {
-  const newProps = Object.assign({}, props, {
-    renderAnswer: props.renderPlayer,
-  });
-  delete newProps.renderPlayer;
-  return newProps;
-};
-
-const ArtistQuestionScreenWrapped = withActivePlayer(
-    withTransformProps(transformPlayerToAnswer)(ArtistQuestionScreen)
-);
-const GenreQuestionScreenWrapped = withUserAnswer(
-    withActivePlayer(
-        withTransformProps(transformPlayerToAnswer)(GenreQuestionScreen))
-);
 
 const Type = {
   ARTIST: `game--artist`,
@@ -36,81 +9,9 @@ const Type = {
 };
 
 class App extends React.PureComponent {
-  _getScreen(questions, step) {
-    const question = questions[step];
-    const {
-      gameOver,
-      onReplayClick
-    } = this.props;
-
-    if (gameOver) {
-      return <LosingScene
-        onClick={() => onReplayClick()}
-      />;
-    }
-
-    if (step >= questions.length) {
-      const {
-        mistakes,
-      } = this.props;
-
-      return <VictoryScene
-        mistakes={mistakes}
-        onClick={() => onReplayClick()}
-      />;
-    } else if (!question) {
-      const {
-        gameTime,
-        maxMistakes,
-        onWelcomeScreenClick,
-      } = this.props;
-
-      return <WelcomeScreen
-        time={gameTime}
-        errorCount={maxMistakes}
-        onClick={onWelcomeScreenClick}
-      />;
-    }
-
-    const {
-      onUserAnswer,
-      mistakes,
-      maxMistakes
-    } = this.props;
-
-    switch (question.type) {
-      case `genre`: return <GenreQuestionScreenWrapped
-        question={question}
-        onAnswer={(userAnswer) => onUserAnswer(
-            userAnswer,
-            questions,
-            step,
-            mistakes,
-            maxMistakes
-        )}
-        key={`genre-question-screen${this.props.step}`}
-      />;
-
-      case `artist`: return <ArtistQuestionScreenWrapped
-        question={question}
-        onAnswer={(userAnswer) => onUserAnswer(
-            userAnswer,
-            questions,
-            step,
-            mistakes,
-            maxMistakes
-        )}
-        key={`artist-question-screen${this.props.step}`}
-      />;
-    }
-
-    return null;
-  }
-
   render() {
     const {
-      questions,
-      step,
+      renderScreen,
       mistakes,
     } = this.props;
 
@@ -135,67 +36,14 @@ class App extends React.PureComponent {
           mistakes={mistakes}
         />}
       </header>
-      {this._getScreen(questions, step)}
+      {renderScreen()}
     </section>;
-
   }
 }
 
 App.propTypes = {
-  gameOver: PropTypes.bool.isRequired,
-  gameTime: PropTypes.number.isRequired,
-  maxMistakes: PropTypes.number.isRequired,
   mistakes: PropTypes.number.isRequired,
-  onReplayClick: PropTypes.func.isRequired,
-  onUserAnswer: PropTypes.func.isRequired,
-  onWelcomeScreenClick: PropTypes.func.isRequired,
-  questions: PropTypes.arrayOf(PropTypes.shape({
-    type: PropTypes.oneOf([`genre`, `artist`]).isRequired,
-    genre: PropTypes.oneOf([`rock`, `pop`, `jazz`]),
-    song: PropTypes.shape({
-      artist: PropTypes.string,
-      src: PropTypes.string,
-    }),
-    answers: PropTypes.arrayOf(PropTypes.shape({
-      genre: PropTypes.oneOf([`rock`, `pop`, `jazz`]),
-      src: PropTypes.string,
-      picture: PropTypes.string,
-      artist: PropTypes.string,
-    })).isRequired,
-  })),
-  step: PropTypes.number.isRequired,
+  renderScreen: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state, ownProps) =>
-  Object.assign({}, ownProps, {
-    step: state.step,
-    mistakes: state.mistakes,
-    gameOver: state.gameOver,
-  });
-
-const mapDispatchToProps = (dispatch) => ({
-  onWelcomeScreenClick: () => {
-    dispatch(ActionCreator.incrementStep());
-  },
-
-  onReplayClick: () => {
-    dispatch(ActionCreator.restart());
-  },
-
-  onUserAnswer: (userAnswer, questions, step, mistakes, maxMistakes) => {
-    dispatch(ActionCreator.incrementStep());
-    dispatch(ActionCreator.incrementMistake(
-        userAnswer,
-        questions[step],
-        mistakes,
-        maxMistakes
-    ));
-  },
-});
-
-export {App};
-
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(App);
+export default App;
